introduce module-structure
